pyproject.toml
[project]
name = "ze-delivery"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-environ>=0.12.0",
    "djangorestframework-gis>=1.2.0",
    "djangorestframework-stubs>=3.16.6",
    "env>=0.1.0",
    "psycopg2-binary>=2.9.11",
]
